requirements.txt
# install torch manuall
Pillow==9.5.0
tqdm==4.65.0
opencv-python==4.8.0.74
matplotlib==3.7.2
captum==0.6.0
scikit-learn==1.3.0
ftfy==6.1.1
regex==2023.6.3
transformers==4.30.2
fastapi==0.100.0
uvicorn==0.23.0
python-multipart==0.0.6
clip_anytorch==2.5.2
gdown==4.7.1